+++ R/ci_profile_boot.R
@@ -59,6 +59,51 @@
 
 
 #' @keywords internal
+.ci_profile_merMod <- function(x, ci, profiled, ...) {
+  out <- stats::confint(profiled, level = ci, ...)
+  rownames(out) <- gsub("`", "", rownames(out), fixed = TRUE)
+  out <- out[rownames(out) %in% insight::find_parameters(x, effects = "fixed")$conditional, ]
+  names(out) <- c("CI_low", "CI_high")
+
+  # Clean up
+  out$Parameter <- row.names(out)
+  out$CI <- ci
+  out <- out[c("Parameter", "CI", "CI_low", "CI_high")]
+  row.names(out) <- NULL
+  out
+}
+
+
+#' @importFrom insight find_parameters
+#' @keywords internal
+.ci_profile_glmmTMB <- function(x, ci, profiled, component, ...) {
+  out <- stats::confint(profiled, level = ci, ...)
+  rownames(out) <- gsub("`", "", rownames(out), fixed = TRUE)
+
+  pars <- insight::find_parameters(x, effects = "fixed")
+
+  param_names <- switch(
+    component,
+    "conditional" = pars$conditional,
+    "zi" = ,
+    "zero_inflated" = paste0("zi~", pars$zero_inflated),
+    c(pars$conditional, paste0("zi~", pars$zero_inflated))
+  )
+
+  out <- out[rownames(out) %in% param_names, ]
+  names(out) <- c("CI_low", "CI_high")
+
+  # Clean up
+  out$Parameter <- row.names(out)
+  out$CI <- ci
+  out <- out[c("Parameter", "CI", "CI_low", "CI_high")]
+  row.names(out) <- NULL
+  out
+}
+
+
+
+#' @keywords internal
 .ci_boot_merMod <- function(x, ci, ...) {
   if (!requireNamespace("lme4", quietly = TRUE)) {
     stop("Package 'lme4' required for this function to work. Please install it by running `install.packages('lme4')`.")