+++ R/methods_lme4.R
@@ -164,7 +164,7 @@ model_parameters.merMod <- function(model,
 #'
 #' @param x A statistical model.
 #' @param ci Confidence Interval (CI) level. Default to 0.95 (95\%).
-#' @param method For mixed models, can be \code{\link[=p_value_wald]{"wald"}} (default), \code{\link[=p_value_ml1]{"ml1"}} or \code{\link[=p_value_betwithin]{"betwithin"}}. For linear mixed model, can also be \code{\link[=p_value_satterthwaite]{"satterthwaite"}}, \code{\link[=p_value_kenward]{"kenward"}} or \code{"boot"} and \code{lme4::confint.merMod}). For (generalized) linear models, can be \code{"robust"} to compute confidence intervals based on robust covariance matrix estimation, and for generalized linear models, may also be \code{"profile"} (default) or \code{"wald"}.
+#' @param method For mixed models, can be \code{\link[=p_value_wald]{"wald"}} (default), \code{\link[=p_value_ml1]{"ml1"}} or \code{\link[=p_value_betwithin]{"betwithin"}}. For linear mixed model, can also be \code{\link[=p_value_satterthwaite]{"satterthwaite"}}, \code{\link[=p_value_kenward]{"kenward"}} or \code{"boot"} (see \code{lme4::confint.merMod}). For (generalized) linear models, can be \code{"robust"} to compute confidence intervals based on robust covariance matrix estimation, and for generalized linear models and models from packages \pkg{lme4} or \pkg{glmmTMB}, may also be \code{"profile"} (default) or \code{"wald"}.
 #' @param ... Arguments passed down to \code{standard_error_robust()} when confidence intervals or p-values based on robust standard errors should be computed.
 #' @inheritParams simulate_model
 #' @inheritParams standard_error
@@ -193,10 +193,11 @@ model_parameters.merMod <- function(model,
 #'   ci(model, component = "zi")
 #' }
 #' }
+#' @importFrom stats profile
 #' @export
 ci.merMod <- function(x,
                       ci = 0.95,
-                      method = c("wald", "ml1", "betwithin", "satterthwaite", "kenward", "boot"),
+                      method = c("wald", "ml1", "betwithin", "satterthwaite", "kenward", "boot", "profile"),
                       ...) {
   method <- tolower(method)
   method <- match.arg(method)
@@ -226,6 +227,12 @@ ci.merMod <- function(x,
     out <- lapply(ci, function(ci, x) .ci_boot_merMod(x, ci, ...), x = x)
     out <- do.call(rbind, out)
     row.names(out) <- NULL
+
+    # profiles CIs
+  } else if (method == "profile") {
+    pp <- stats::profile(x)
+    out <- lapply(ci, function(i) .ci_profile_merMod(model = x, ci = i, profiled = pp, ...))
+    out <- do.call(rbind, out)
   }
 
   out